✨ feat: enhance email verification configuration and logic
- Updated the authentication configuration to include a new 'verification' section for email link expiration settings.
- Ensured that the email verification expiration time is consistently cast to an integer across relevant classes.
- Modified the VerifyEmailChangeRequest, EmailChangeVerification, and UserRepository to utilize the updated configuration for improved clarity and reliability in email change verification processes.

+++ stubs/app/Http/Requests/VerifyEmailChangeRequest.php
@@ -37,7 +37,7 @@ class VerifyEmailChangeRequest extends FormRequest
         }
 
         // Check if the token hasn't expired (default 60 minutes)
-        $expireTime = config('auth.verification.expire', 60);
+        $expireTime = (int) config('auth.verification.expire', 60);
         if ($user->email_change_requested_at->addMinutes($expireTime)->isPast()) {
             return false;
         }
